+++ src/App.css
@@ -64,6 +64,10 @@
   margin-left: 10px;
 }
 
+.margin-right {
+  margin-right: 10px;
+}
+
 /* 
     All pages
 */
